버그 수정(테스트 코드 오류)
- transactionInput 테스트 코드 에러 안나도록 수정 ( 코드 리팩토링 )
