+++ src/transactionDetail/CategoryContainer.jsx
@@ -1,0 +1,70 @@
+import React from 'react';
+
+import { useDispatch, useSelector } from 'react-redux';
+
+import styled from '@emotion/styled';
+import mediaquery from '../style/mediaquery';
+
+import { get } from '../utils/utils';
+
+import {
+  selectCategory,
+  changeTransactionCategory,
+} from '../slice';
+
+import ExpenseCategory from './ExpenseCategory';
+import IncomeCategory from './IncomeCategory';
+
+const OptionBox = styled.div(mediaquery({
+  width: '100%',
+  height: ['18%', '18%', '19%', '15%', '15%'],
+  // height: ['15%', '17%', '15%', '16%', '16%'],
+  // height: ['50%', '50%', '50%', '47.5%', '47.5%'],
+  marginBottom: ['0', '0', '0', '.5em', '.5em'],
+  // padding: '.5em',
+  textAlign: 'center',
+  lineHeight: ['2.2em', '2.4em', '2.2em', '4em', '4em'],
+  '& label': {
+    float: 'left',
+    width: '25%',
+    height: '100%',
+    fontSize: ['.6em', '.6em', '.7em', '.7em', '.7em'],
+  },
+}));
+
+export default function CategoryContainer() {
+  const dispatch = useDispatch();
+
+  const selectedType = useSelector(get('selectedType'));
+  const transaction = useSelector(get('transaction'));
+
+  const handleChange = (name) => {
+    dispatch(selectCategory(name));
+    dispatch(changeTransactionCategory(name));
+  };
+
+  return (
+    <OptionBox>
+      <label htmlFor="category">
+        카테고리
+      </label>
+      <div>
+        {
+          selectedType === '지출'
+            ? (
+              <ExpenseCategory
+                transaction={transaction}
+                onChange={handleChange}
+              />
+            )
+            : (
+              <IncomeCategory
+                transaction={transaction}
+                onChange={handleChange}
+              />
+            )
+        }
+      </div>
+    </OptionBox>
+  );
+}
