+++ src/transactionDetail/SubmitContainer.jsx
@@ -1,12 +1,12 @@
 import { useDispatch, useSelector } from 'react-redux';
 
-import TransactionInput from './TransactionInput';
+import styled from '@emotion/styled';
+import colors from '../style/colors';
+import mediaquery from '../style/mediaquery';
 
 import { get } from '../utils/utils';
 
 import {
-  changeBreakdownFields,
-  changeTransactionFields,
   clearTransactionFields,
   setTransaction,
   addMonthlyTransaction,
@@ -14,19 +14,26 @@ import {
   clearTargetId,
 } from '../slice';
 
-export default function TransactionInputContainer() {
+const SubmitBox = styled.div(mediaquery({
+  width: ['30%', '30%', '30%', '50%', '50%'],
+  height: ['9%', '9%', '9%', '7%', '7%'],
+  margin: ['0.5em auto 0', '0.5em auto 0', '0.5em auto 0', '3em auto 0', '3em auto 0'],
+  textAlign: 'center',
+  lineHeight: '2em',
+  borderRadius: '.2em',
+  color: `${colors.white}`,
+  backgroundColor: `${colors.teal}`,
+  fontSize: ['.7em', '.7em', '.8em', '.8em', '.8em'],
+  fontWeight: '600',
+  letterSpacing: 5,
+}));
+
+export default function SubmitContainer() {
   const dispatch = useDispatch();
   const transaction = useSelector(get('transaction'));
   const targetId = useSelector(get('targetId'));
+
   const { transactionFields } = transaction;
-
-  const handleChangeBreakdown = ({ value }) => {
-    dispatch(changeBreakdownFields({ value }));
-  };
-
-  const handleChangeFields = ({ name, value }) => {
-    dispatch(changeTransactionFields({ name, value }));
-  };
 
   const handleSubmit = (id) => {
     if (transactionFields.breakdown === 0) {
@@ -59,14 +66,13 @@ export default function TransactionInputContainer() {
   };
 
   return (
-    <>
-      <TransactionInput
-        targetId={targetId}
-        fields={transactionFields}
-        onChange={handleChangeFields}
-        onChangeBreakdown={handleChangeBreakdown}
-        onClick={handleSubmit}
-      />
-    </>
+    <SubmitBox>
+      <button
+        type="button"
+        onClick={() => handleSubmit(targetId)}
+      >
+        저장
+      </button>
+    </SubmitBox>
   );
 }
